Move gh-pages to docs/
The files in the root directory are the gem or repository files only.
All other files related to GitHub Pages hosting are now moved to a
separate directory "docs/". The directory structure was refactored to
match the new changes.

docs/ contains two configuration files:

 - _config.yml     : used for GitHub Pages hosting.
 - _config.dev.yml : used for local development.

To run the theme locally use the following commands:

    gem build portfolyou-jekyll-theme.gemspec
    cd docs/
    bundle install
    bundle exec jekyll serve --config _config.dev.yml